strip out shared average rating calculation per team request
